LC-maximum-size-subarray-sum-equals-k [Runtime: 158 ms; Memory: 131 MB]

+++ hash-table/maximum-size-subarray-sum-equals-k.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int maxSubArrayLen(vector<int>& nums, int k) {
 
-        int maxLen = 0;
-        unordered_map<int, int> prefixIdx;
-        int prefixSum = 0;
+        long long maxLen = 0;
+        unordered_map<long long, long long> prefixIdx;
+        long long prefixSum = 0;
 
         for(int i=0; i<nums.size(); i++){
             prefixSum += nums[i];
